requirements.txt
psycopg2-binary
MLB-Statsapi
pybaseball
python-dotenv
pandas
numpy
requests